requirements.txt
# Core dependencies (required)
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
qrcode[pil]==7.4.2
Pillow==12.1.0
python-dotenv==1.0.0

# Face recognition dependencies (using OpenCV for flexible matching)
opencv-python>=4.9.0
numpy>=2.0.0
face-recognition>=1.3.0